+++ pkg/image/importer/importer.go
@@ -1,29 +1,19 @@
 package importer
 
 import (
-	"encoding/json"
 	"fmt"
-	"net/http"
 	"net/url"
-	"path"
 	"strings"
-	"time"
 
 	"github.com/golang/glog"
 	gocontext "golang.org/x/net/context"
 
 	"github.com/docker/distribution"
-	"github.com/docker/distribution/context"
 	"github.com/docker/distribution/digest"
-	"github.com/docker/distribution/manifest/schema1"
 	"github.com/docker/distribution/reference"
 	"github.com/docker/distribution/registry/api/errcode"
 	"github.com/docker/distribution/registry/api/v2"
-	registryclient "github.com/docker/distribution/registry/client"
-	"github.com/docker/distribution/registry/client/auth"
-	"github.com/docker/distribution/registry/client/transport"
 
-	kapi "k8s.io/kubernetes/pkg/api"
 	kapierrors "k8s.io/kubernetes/pkg/api/errors"
 	"k8s.io/kubernetes/pkg/api/unversioned"
 	"k8s.io/kubernetes/pkg/util"
@@ -32,7 +22,6 @@ import (
 
 	"github.com/openshift/origin/pkg/dockerregistry"
 	"github.com/openshift/origin/pkg/image/api"
-	"github.com/openshift/origin/pkg/image/api/dockerpre012"
 )
 
 // Add a dockerregistry.Client to the passed context with this key to support v1 Docker registry importing
@@ -48,15 +37,6 @@ type RepositoryRetriever interface {
 	// Repository returns a properly authenticated distribution.Repository for the given registry, repository
 	// name, and insecure toleration behavior.
 	Repository(ctx gocontext.Context, registry *url.URL, repoName string, insecure bool) (distribution.Repository, error)
-}
-
-// ErrNotV2Registry is returned when the server does not report itself as a V2 Docker registry
-type ErrNotV2Registry struct {
-	Registry string
-}
-
-func (e *ErrNotV2Registry) Error() string {
-	return fmt.Sprintf("endpoint %q does not support v2 API", e.Registry)
 }
 
 // ImageStreamImport implements an import strategy for Docker images. It keeps a cache of images
@@ -604,199 +584,4 @@ func setImageImportStatus(images *api.ImageStreamImport, i int, err error) {
 
 func invalidStatus(position string, errs ...*field.Error) unversioned.Status {
 	return kapierrors.NewInvalid(api.Kind(""), position, errs).(kapierrors.APIStatus).Status()
-}
-
-// NewContext is capable of creating RepositoryRetrievers.
-func NewContext(transport, insecureTransport http.RoundTripper) Context {
-	return Context{
-		Transport:         transport,
-		InsecureTransport: insecureTransport,
-		Challenges:        auth.NewSimpleChallengeManager(),
-	}
-}
-
-type Context struct {
-	Transport         http.RoundTripper
-	InsecureTransport http.RoundTripper
-	Challenges        auth.ChallengeManager
-}
-
-func (c Context) WithCredentials(credentials auth.CredentialStore) RepositoryRetriever {
-	return &repositoryRetriever{
-		context:     c,
-		credentials: credentials,
-
-		pings:    make(map[url.URL]error),
-		redirect: make(map[url.URL]*url.URL),
-	}
-}
-
-type repositoryRetriever struct {
-	context     Context
-	credentials auth.CredentialStore
-
-	pings    map[url.URL]error
-	redirect map[url.URL]*url.URL
-}
-
-func (r *repositoryRetriever) Repository(ctx gocontext.Context, registry *url.URL, repoName string, insecure bool) (distribution.Repository, error) {
-	t := r.context.Transport
-	if insecure && r.context.InsecureTransport != nil {
-		t = r.context.InsecureTransport
-	}
-	src := *registry
-	// ping the registry to get challenge headers
-	if err, ok := r.pings[src]; ok {
-		if err != nil {
-			return nil, err
-		}
-		if redirect, ok := r.redirect[src]; ok {
-			src = *redirect
-		}
-	} else {
-		redirect, err := r.ping(src, insecure, t)
-		r.pings[src] = err
-		if err != nil {
-			return nil, err
-		}
-		if redirect != nil {
-			r.redirect[src] = redirect
-			src = *redirect
-		}
-	}
-
-	rt := transport.NewTransport(
-		t,
-		// TODO: slightly smarter authorizer that retries unauthenticated requests
-		// TODO: make multiple attempts if the first credential fails
-		auth.NewAuthorizer(
-			r.context.Challenges,
-			auth.NewTokenHandler(t, r.credentials, repoName, "pull"),
-			auth.NewBasicHandler(r.credentials),
-		),
-	)
-	return registryclient.NewRepository(context.Context(ctx), repoName, src.String(), rt)
-}
-
-func (r *repositoryRetriever) ping(registry url.URL, insecure bool, transport http.RoundTripper) (*url.URL, error) {
-	pingClient := &http.Client{
-		Transport: transport,
-		Timeout:   15 * time.Second,
-	}
-	target := registry
-	target.Path = path.Join(target.Path, "v2") + "/"
-	req, err := http.NewRequest("GET", target.String(), nil)
-	if err != nil {
-		return nil, err
-	}
-	resp, err := pingClient.Do(req)
-	if err != nil {
-		if insecure && registry.Scheme == "https" {
-			glog.V(5).Infof("Falling back to an HTTP check for an insecure registry %s: %v", registry, err)
-			registry.Scheme = "http"
-			_, nErr := r.ping(registry, true, transport)
-			if nErr != nil {
-				return nil, nErr
-			}
-			return &registry, nil
-		}
-		return nil, err
-	}
-	defer resp.Body.Close()
-
-	versions := auth.APIVersions(resp, "Docker-Distribution-API-Version")
-	if len(versions) == 0 {
-		glog.V(5).Infof("Registry responded to v2 Docker endpoint, but has no header for Docker Distribution %s: %d, %#v", req.URL, resp.StatusCode, resp.Header)
-		return nil, &ErrNotV2Registry{Registry: registry.String()}
-	}
-
-	r.context.Challenges.AddResponse(resp)
-
-	return nil, nil
-}
-
-func schema1ToImage(manifest *schema1.SignedManifest, d digest.Digest) (*api.Image, error) {
-	if len(manifest.History) == 0 {
-		return nil, fmt.Errorf("image has no v1Compatibility history and cannot be used")
-	}
-	dockerImage, err := unmarshalDockerImage([]byte(manifest.History[0].V1Compatibility))
-	if err != nil {
-		return nil, err
-	}
-	if len(d) > 0 {
-		dockerImage.ID = d.String()
-	} else {
-		if p, err := manifest.Payload(); err == nil {
-			d, err := digest.FromBytes(p)
-			if err != nil {
-				return nil, fmt.Errorf("unable to create digest from image payload: %v", err)
-			}
-			dockerImage.ID = d.String()
-		} else {
-			d, err := digest.FromBytes(manifest.Raw)
-			if err != nil {
-				return nil, fmt.Errorf("unable to create digest from image bytes: %v", err)
-			}
-			dockerImage.ID = d.String()
-		}
-	}
-	image := &api.Image{
-		ObjectMeta: kapi.ObjectMeta{
-			Name: dockerImage.ID,
-		},
-		DockerImageMetadata:        *dockerImage,
-		DockerImageManifest:        string(manifest.Raw),
-		DockerImageMetadataVersion: "1.0",
-	}
-
-	return image, nil
-}
-
-func schema0ToImage(dockerImage *dockerregistry.Image, id string) (*api.Image, error) {
-	var baseImage api.DockerImage
-	if err := kapi.Scheme.Convert(&dockerImage.Image, &baseImage); err != nil {
-		return nil, fmt.Errorf("could not convert image: %#v", err)
-	}
-
-	image := &api.Image{
-		ObjectMeta: kapi.ObjectMeta{
-			Name: dockerImage.ID,
-		},
-		DockerImageMetadata:        baseImage,
-		DockerImageMetadataVersion: "1.0",
-	}
-
-	return image, nil
-}
-
-func unmarshalDockerImage(body []byte) (*api.DockerImage, error) {
-	var image dockerpre012.DockerImage
-	if err := json.Unmarshal(body, &image); err != nil {
-		return nil, err
-	}
-	dockerImage := &api.DockerImage{}
-	if err := kapi.Scheme.Convert(&image, dockerImage); err != nil {
-		return nil, err
-	}
-	return dockerImage, nil
-}
-
-func isDockerError(err error, code errcode.ErrorCode) bool {
-	switch t := err.(type) {
-	case errcode.Errors:
-		for _, err := range t {
-			if isDockerError(err, code) {
-				return true
-			}
-		}
-	case errcode.ErrorCode:
-		if code == t {
-			return true
-		}
-	case errcode.Error:
-		if t.ErrorCode() == code {
-			return true
-		}
-	}
-	return false
 }
